split tests for logging
