+++ python/ray/serve/tests/test_logging_2.py
@@ -1,0 +1,97 @@
+import os
+import re
+import uuid
+
+import httpx
+import pytest
+from fastapi import FastAPI
+
+import ray
+from ray import serve
+from ray._common.test_utils import wait_for_condition
+from ray.serve._private.logging_utils import (
+    get_serve_logs_dir,
+)
+from ray.serve._private.utils import get_component_file_name
+from ray.util.state import list_actors, list_nodes
+
+
+@pytest.fixture
+def serve_and_ray_shutdown():
+    serve.shutdown()
+    ray.shutdown()
+    yield
+
+
+def check_log_file(log_file: str, expected_regex: list, check_contains: bool = True):
+    with open(log_file, "r") as f:
+        s = f.read()
+        for regex in expected_regex:
+            if check_contains:
+                assert re.findall(regex, s) != []
+            else:
+                assert re.findall(regex, s) == []
+
+
+def test_http_access_log_in_proxy_logs_file(serve_instance):
+    name = "deployment_name"
+    fastapi_app = FastAPI()
+
+    @serve.deployment(name=name)
+    @serve.ingress(fastapi_app)
+    class Handler:
+        @fastapi_app.get("/")
+        def get_root(self):
+            return "Hello World!"
+
+    serve.run(Handler.bind(), logging_config={"encoding": "TEXT"})
+
+    # Get log file information
+    nodes = list_nodes()
+    serve_log_dir = get_serve_logs_dir()
+    node_ip_address = nodes[0].node_ip
+    proxy_log_file_name = get_component_file_name(
+        "proxy", node_ip_address, component_type=None, suffix=".log"
+    )
+    proxy_log_path = os.path.join(serve_log_dir, proxy_log_file_name)
+
+    request_id = str(uuid.uuid4())
+    response = httpx.get("http://localhost:8000", headers={"X-Request-ID": request_id})
+    assert response.status_code == 200
+
+    def verify_request_id_in_logs(proxy_log_path, request_id):
+        with open(proxy_log_path, "r") as f:
+            for line in f:
+                if request_id in line:
+                    return True
+        return False
+
+    wait_for_condition(
+        verify_request_id_in_logs, proxy_log_path=proxy_log_path, request_id=request_id
+    )
+
+
+def test_start_serve_with_logging_config(self, serve_and_ray_shutdown):
+    serve.start(logging_config={"log_level": "DEBUG", "encoding": "JSON"})
+    serve_log_dir = get_serve_logs_dir()
+    # Check controller log
+    actors = list_actors()
+    expected_log_regex = [".*logger with logging config.*"]
+    for actor in actors:
+        print(actor["name"])
+        if "SERVE_CONTROLLER_ACTOR" == actor["name"]:
+            controller_pid = actor["pid"]
+    controller_log_file_name = get_component_file_name(
+        "controller", controller_pid, component_type=None, suffix=".log"
+    )
+    controller_log_path = os.path.join(serve_log_dir, controller_log_file_name)
+    check_log_file(controller_log_path, expected_log_regex)
+
+    # Check proxy log
+    nodes = list_nodes()
+    node_ip_address = nodes[0].node_ip
+    proxy_log_file_name = get_component_file_name(
+        "proxy", node_ip_address, component_type=None, suffix=".log"
+    )
+    proxy_log_path = os.path.join(serve_log_dir, proxy_log_file_name)
+    check_log_file(proxy_log_path, expected_log_regex)
